Post a json object

+++ src/main/java/com/example/demo/model/Category.java
@@ -2,20 +2,20 @@ package com.example.demo.model;
 
 public class Category {
 
-    private Long id;
+    private int id;
     private String name;
 
-    public Category(Long id, String name) {
+    public Category(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
